[Documentation update] Memory Management [Deodhai]: Bugs fixed in popup window. Now popup window is also a normal window with POPUP flag and little different behaviour than normal window [Chitralekha]: Bug fixes in popup menu

+++ Kernel/Mm/pmmngr.cpp
@@ -116,7 +116,7 @@ void AuPmmngrLockPage(uint64_t Address) {
 void AuPmmngrLockPages(void *Address, size_t Size) {
 	uint64_t addr = (uint64_t)Address;
 	for (int i = 0; i < Size; i++)
-		AuPmmngrLockPage(addr + i * 4096);
+		AuPmmngrLockPage(addr + static_cast<uint64_t>(i) * 4096);
 }
 
 /*
